# src/flask_app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sistem Planificare Examene API (Flask)",
    "description": "API pentru sistemul de planificare examene FIESC implementat în Flask",
    "version": "1.0.0",
    "contact": {
      "email": "planificare@fiesc.usv.ro"
    }
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
...
  "paths": {
...
    "/schedules": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "Obținerea planificărilor",
        "description": "Endpoint pentru obținerea planificărilor",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "query",
            "description": "ID-ul grupei",
            "required": false,
            "type": "integer"
          },
          {
            "name": "teacher_id",
            "in": "query",
            "description": "ID-ul cadrului didactic",
            "required": false,
            "type": "integer"
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "Data de început pentru filtrare (format: YYYY-MM-DD)",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "Data de sfârșit pentru filtrare (format: YYYY-MM-DD)",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Statusul planificării",
            "required": false,
            "type": "string",
            "enum": [
              "proposed",
              "approved",
              "rejected"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Planificări obținute cu succes",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Schedule"
              }
            }
          },
          "401": {
            "description": "Utilizator neautentificat"
          },
          "500": {
            "description": "Eroare la obținerea planificărilor"
          }
        }
      },
      "post": {
        "tags": [
          "schedules"
        ],
        "summary": "Crearea unei planificări",
        "description": "Endpoint pentru crearea unei planificări",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Datele planificării",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ScheduleCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Planificare creată cu succes",
            "schema": {
              "$ref": "#/definitions/Schedule"
            }
          },
          "400": {
            "description": "Date lipsă sau format dată/timp invalid"
          },
          "401": {
            "description": "Utilizator neautentificat"
          },
          "403": {
            "description": "Utilizator fără permisiuni"
          },
          "500": {
            "description": "Eroare la crearea planificării"
          }
        }
      }
    },
    "/schedules/{schedule_id}": {
      "get": {
        "tags": [
          "schedules"
        ],
        "summary": "Obținerea unei planificări după ID",
        "description": "Endpoint pentru obținerea unei planificări după ID",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "description": "ID-ul planificării",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Planificare obținută cu succes",
            "schema": {
              "$ref": "#/definitions/Schedule"
            }
          },
          "401": {
            "description": "Utilizator neautentificat"
          },
          "404": {
            "description": "Planificare negăsită"
          }
        }
      },
      "put": {
        "tags": [
          "schedules"
        ],
        "summary": "Actualizarea unei planificări",
        "description": "Endpoint pentru actualizarea unei planificări",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "description": "ID-ul planificării",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Datele planificării",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ScheduleUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Planificare actualizată cu succes",
            "schema": {
              "$ref": "#/definitions/Schedule"
            }
          },
          "400": {
            "description": "Date lipsă sau format dată/timp invalid"
          },
          "401": {
            "description": "Utilizator neautentificat"
          },
          "403": {
            "description": "Utilizator fără permisiuni"
          },
          "404": {
            "description": "Planificare negăsită sau eroare la actualizare"
          }
        }
      },
      "delete": {
        "tags": [
          "schedules"
        ],
        "summary": "Ștergerea unei planificări",
        "description": "Endpoint pentru ștergerea unei planificări",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "description": "ID-ul planificării",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Planificare ștearsă cu succes",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Planificare ștearsă cu succes"
                }
              }
            }
          },
          "401": {
            "description": "Utilizator neautentificat"
          },
          "403": {
            "description": "Utilizator fără permisiuni"
          },
          "404": {
            "description": "Planificare negăsită sau eroare la ștergere"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Subject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Programare Web"
        },
        "acronym": {
          "type": "string",
          "example": "PW"
        }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "firstName": {
          "type": "string",
          "example": "Nume"
        },
        "lastName": {
          "type": "string",
          "example": "Prenume"
        },
        "email": {
          "type": "string",
          "example": "email@usv.ro"
        }
      }
    },
    "Group": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "3A4"
        },
        "year": {
          "type": "integer",
          "example": 3
        },
        "specialization": {
          "type": "string",
          "example": "Calculatoare"
        }
      }
    },
    "Room": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "C201"
        },
        "capacity": {
          "type": "integer",
          "example": 30
        },
        "building": {
          "type": "string",
          "example": "C"
        }
      }
    },
    "Schedule": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "subject": {
          "$ref": "#/definitions/Subject"
        },
        "teacher": {
          "$ref": "#/definitions/Teacher"
        },
        "group": {
          "$ref": "#/definitions/Group"
        },
        "room": {
          "$ref": "#/definitions/Room"
        },
        "date": {
          "type": "string",
          "format": "date",
          "example": "2023-06-15"
        },
        "startTime": {
          "type": "string",
          "format": "time",
          "example": "10:00"
        },
        "endTime": {
          "type": "string",
          "format": "time",
          "example": "12:00"
        },
        "status": {
          "type": "string",
          "enum": [
            "proposed",
            "approved",
            "rejected"
          ],
          "example": "approved"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "example": "2023-05-01T12:00:00Z"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "example": "2023-05-02T14:30:00Z"
        }
      }
    },
    "ScheduleCreate": {
      "type": "object",
      "required": [
        "subjectId",
        "teacherId",
        "groupId",
        "date",
        "startTime",
        "endTime"
      ],
      "properties": {
        "subjectId": {
          "type": "integer",
          "example": 1
        },
        "teacherId": {
          "type": "integer",
          "example": 1
        },
        "groupId": {
          "type": "integer",
          "example": 1
        },
        "roomId": {
          "type": "integer",
          "example": 1
        },
        "date": {
          "type": "string",
          "format": "date",
          "example": "2023-06-15"
        },
        "startTime": {
          "type": "string",
          "format": "time",
          "example": "10:00"
        },
        "endTime": {
          "type": "string",
          "format": "time",
          "example": "12:00"
        }
      }
    },
    "ScheduleUpdate": {
      "type": "object",
      "properties": {
        "subjectId": {
          "type": "integer",
          "example": 1
        },
        "teacherId": {
          "type": "integer",
          "example": 1
        },
        "groupId": {
          "type": "integer",
          "example": 1
        },
        "roomId": {
          "type": "integer",
          "example": 1
        },
        "date": {
          "type": "string",
          "format": "date",
          "example": "2023-06-15"
        },
        "startTime": {
          "type": "string",
          "format": "time",
          "example": "10:00"
        },
        "endTime": {
          "type": "string",
          "format": "time",
          "example": "12:00"
        },
        "status": {
          "type": "string",
          "enum": [
            "proposed",
            "approved",
            "rejected"
          ],
          "example": "approved"
        }
      }
    },
...
  }
}